Initial Push
Adding a simple menu toggle to allow the server to start with a full wildcard.

This isn't ideal - it should be using 0.0.255.255 for tighter security, , but it doesn't seem to work that way.

Either way, these 2 minor file changes allow the possibility of debugging things on local networks - such as from a cellphone.